OPENSEARCH/requirements.txt
opensearch-py>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

